Adds types file
